requirements.txt
# Python dependencies for Federated SSL

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pillow>=9.5.0

# Federated learning
flwr>=1.3.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.2.0
pydicom>=2.3.0

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
